+++ chrome/android/java/src/org/chromium/chrome/browser/payments/ui/PaymentAppComparator.java
@@ -58,7 +58,8 @@ import java.util.Comparator;
      */
     private static double getRankingScore(int count, long date) {
         long currentTime = System.currentTimeMillis();
-        return -Math.log((currentTime - date) / (24 * 60 * 60 * 1000) + 2) / Math.log(count + 2);
+        return -Math.log((double) ((currentTime - date) / (24 * 60 * 60 * 1000) + 2))
+                / Math.log(count + 2);
     }
 
     /**
